Move preflight issue extractor to submit_ce/ui/preflight [SUBMISSION-210] David
